ZProjekt/requirements.txt
Flask>=3.0.0
msal>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.9
xhtml2pdf>=0.2.15
celery>=5.3.0
redis>=5.0.0
gunicorn>=21.2.0
